% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_icasa.R
\name{create_icasa}
\alias{create_icasa}
\title{Create a DSSAT ICASA File from NASA - POWER Data}
\usage{
create_icasa(lonlat = NULL, dates = NULL, dsn = NULL,
  file_out = NULL)
}
\arguments{
\item{lonlat}{A numeric vector of geographic coordinates for a cell or region
entered as x, y coordinates.  See argument details for more.}

\item{dates}{A character vector of start and end dates in that order,\cr
\emph{e.g.}, \code{dates = c("1983-01-01", "2017-12-31")}.  See argument details for
more.}

\item{dsn}{A file path where the resulting text file should be stored.}

\item{file_out}{A file name for the resulting text file, \emph{e.g.} "ICASA.txt".
A ".txt" extension will be appended if not or otherwise specified by user.}
}
\description{
Get NASA-POWER values for a single point or region and create
an ICASA format text file suitable for use in DSSAT for crop modelling.
}
\details{
This function is essentially a wrapper for \code{\link{get_power}}
queryies the POWER API and writes a DSSAT ICASA weather file to disk. All
necessary \code{pars} are automatically included in the query.

Further details for each of the arguments are provided in their
respective sections following below.
}
\section{Argument details for \code{lonlat}}{

\describe{
\item{For a single point}{To get a specific cell, 1/2 x 1/2 degree, supply a
length-2 numeric vector giving the decimal degree longitude and latitude in
that order for data to download,\cr
\emph{e.g.}, \code{lonlat = c(151.81, -27.48)}.}

\item{For regional coverage}{To get a region, supply a length-4 numeric
vector as lower left (lat, lon) and upper right (lat, lon) coordinates,
\emph{e.g.}, \code{lonlat = c(ymin, xmin, ymax, xmax)} in that order for a given
region, \emph{e.g.}, a bounding box for the southwestern corner of Australia:
\code{lonlat = c(112.5, -55.5, 115.5, -50.5)}. \emph{Max bounding box is 10 x 10
degrees} of 1/2 x 1/2 degree data, \emph{i.e.}, 100 points maximum in total.}
}
}

\section{Argument details for \code{dates}}{
 If \code{dates} is unspecified, defaults to
a start date of 1983-01-01 (the earliest available data) and an end date of
current date according to the system.

If one date only is provided, it will be treated as both the start date and
the end date and only a single day's values will be returned.
}

\examples{
# Create an ICASA file for Kingsthorpe, Qld from 1985-01-01 to 1985-06-30 and
# save it in "Documents".

\dontrun{
create_icasa(lonlat = c(151.81, -27.48),
             dates = c("1985-01-01", "1985-12-31"),
             dsn = "~/Documents",
             file = "ICASA.txt"
)
}

}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
