% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_met.R
\name{create_met}
\alias{create_met}
\title{Create an APSIM weather.met File from NASA - POWER Data}
\usage{
create_met(latlon = NULL, dates = NULL)
}
\arguments{
\item{latlon}{A numeric vector of geographic coordinates for a cell or region
entered as x, y coordinates.  See argument details for more.}

\item{dates}{A character vector of start and end dates in that order,\cr
\emph{e.g.}, \code{dates = c("1983-01-01", "2017-12-31")}.  See argument details for
more.}
}
\description{
Get NASA-POWER values for a single point or region and create
an APSIM weather.met object suitable for use in APSIM for crop modelling.
}
\details{
This function is essentially a wrapper for \code{\link{get_power}}
and \code{\link[APSIM]{prepareMet}} that simplifies the querying of the
POWER API and returns a .met object.

The weather values from POWER for temperature are 2 metre max and min
temperatures, T2M_MAX and T2M_MIN; radation, ALLSKY_SFC_SW_DWN; and rain,
PRECTOT from the POWER AG community on a daily time-step.

Further details for each of the arguments are provided in their
respective sections following below.
}
\section{Argument details for \code{latlon}}{

\describe{
\item{For a single point}{To get a specific cell, 1/2 x 1/2 degree, supply a
length-2 numeric vector giving the decimal degree longitude and latitude in
that order for data to download,\cr
\emph{e.g.}, \code{latlon = c(-27.48, 151.81)}.}

\item{For regional coverage}{To get a region, supply a length-4 numeric
vector as lower left (lat, lon) and upper right (lat, lon) coordinates,
\emph{e.g.}, \code{latlon = c(ymin, xmin, ymax, xmax)} in that order for a given
region, \emph{e.g.}, a bounding box for the southwestern corner of Australia:
\code{latlon = c(-55.5, 112.5, -50.5, 115.5)}. \emph{Max bounding box is 10 x 10
degrees} of 1/2 x 1/2 degree data, \emph{i.e.}, 100 points maximum in total.}
}
}

\section{Argument details for \code{dates}}{
 If \code{dates} is unspecified, defaults to
a start date of 1983-01-01 (the earliest available data) and an end date of
current date according to the system.

If one date only is provided, it will be treated as both the start date and
the end date and only a single day's values will be returned.
}

\examples{
# Create a .met object for Kingsthorpe, Qld from 1985-01-01 to 1985-06-30.

\dontrun{
Kingsthorpe <- create_met(latlon = c(-27.48, 151.81),
                          dates = c("1985-01-01", "1985-12-31")
)
}

}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
