% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power.R
\name{.check_dates}
\alias{.check_dates}
\title{Check dates for validity when querying the API}
\usage{
.check_dates(dates, lonlat, temporal_api)
}
\arguments{
\item{dates}{User entered \code{dates} value.}

\item{lonlat}{User entered \code{lonlat} value.}

\item{temporal_api}{User entered \code{temporal_api} value.}
}
\value{
Validated dates in a list for use in \code{.build_query}.
}
\description{
Validates user entered dates against \code{lonlat} and \code{temporal_api} values
}
\keyword{internal}
