% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power.R
\name{.check_inputs}
\alias{.check_inputs}
\title{Check User Inputs for get_power for Validity}
\usage{
.check_inputs(
  community,
  dates,
  lonlat,
  pars,
  site_elevation,
  temporal_api,
  wind_elevation,
  wind_surface
)
}
\arguments{
\item{community}{A case-insensitive character vector providing community name:
\dQuote{AG}, \dQuote{RE} or \dQuote{SB}.  See argument details for more.}

\item{lonlat}{A numeric vector of geographic coordinates for a cell or region
entered as x, y (longitude, latitude) coordinates.  See argument details
for more.}

\item{pars}{case-insensitive character vector of solar, meteorological or
climatology parameters to download.  When requesting a single point of x, y
coordinates, a maximum of twenty (20) \code{pars} can be specified at one time,
for \dQuote{daily}, \dQuote{monthly} and \dQuote{climatology}
\code{temporal_api}s.  If the \code{temporal_api} is specified as \dQuote{hourly}
only 15 \code{pars} can be specified in a single query.  See \code{temporal_api} for
more.  These values are checked internally for validity before sending the
query to the \acronym{POWER} \acronym{API}.}

\item{site_elevation}{A user-supplied value for elevation at a single point
in metres.  If provided this will return a corrected atmospheric pressure
value adjusted to the elevation provided.  Only used with \code{lonlat} as a
single point of x, y coordinates, not for use with \dQuote{global} or with
a regional request.}

\item{temporal_api}{A case-insensitive character vector providing the temporal
\acronym{API} end-point for data being queried, supported values are
\dQuote{hourly}, \dQuote{daily}, \dQuote{monthly} or \dQuote{climatology}.
Defaults to \dQuote{daily}.  See argument details for more.}

\item{wind_elevation}{A user-supplied value for elevation at a single point
in metres.  Wind Elevation values in Meters are required to be between 10m
and 300m.  Only used with \code{lonlat} as a single point of x, y coordinates,
not for use with \dQuote{global} or with a regional request.  If this
parameter is provided, the \code{wind_surface} parameter is required with the
request, see
\url{https://power.larc.nasa.gov/docs/methodology/meteorology/wind/}.}

\item{wind_surface}{A user-supplied wind surface for which the corrected
wind-speed is to be supplied.  See \code{wind-surface} section for more detail.}
}
\description{
Check User Inputs for get_power for Validity
}
