% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.check_pars}
\alias{.check_pars}
\title{Check Pars for Validity When Querying API}
\usage{
.check_pars(pars, community, temporal_api)
}
\arguments{
\item{pars}{User entered \code{pars} value.}

\item{community}{User entered \code{community} value.}

\item{temporal_api}{User entered \code{temporal_api} value.}
}
\value{
Validated a collapsed string of  \code{pars} for use in \link{.build_query}.
}
\description{
Validates user entered \code{pars} values against \code{temporal_api} values.
}
\keyword{internal}
