% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power.R
\name{.format_dates_re_sb}
\alias{.format_dates_re_sb}
\title{Format date columns in POWER data frame for the re community}
\usage{
.format_dates_re_sb(power_response)
}
\arguments{
\item{power_response}{A tidy data.frame resulting from \code{\link[=.build_query]{.build_query()}}.}
}
\value{
A tidy data frame of 'POWER' data with additional date information
columns.
}
\description{
Formats columns as integers for DOY and adds columns for year, month and day.
}
\keyword{internal}
