% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_request_options.R
\name{.get_max_tries}
\alias{.get_max_tries}
\title{Get Max Tries Option}
\usage{
.get_max_tries()
}
\value{
An integer giving the maximum number of attempts.
}
\description{
Retrieves the maximum number of request attempts for POWER API requests
from the \code{nasapower.max_tries} option, falling back to a default of \code{6L}
if not set.
}
\examples{
# Use default
.get_max_tries()

# Override via option
options(nasapower.max_tries = 3L)
.get_max_tries()

}
\keyword{internal}
