% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_request_options.R
\name{.get_timeout_connect}
\alias{.get_timeout_connect}
\title{Get Connection Timeout Option}
\usage{
.get_timeout_connect()
}
\value{
An integer giving the connection timeout in seconds.
}
\description{
Retrieves the connection timeout value for POWER API requests from the
\code{nasapower.timeout.connect} option, falling back to a default of \code{5L}
seconds if not set.
}
\examples{
# Use default
.get_timeout_connect()

# Override via option
options(nasapower.timeout.connect = 10L)
.get_timeout_connect()

}
\keyword{internal}
