% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.handle_http_response}
\alias{.handle_http_response}
\title{Handle HTTP Response Errors}
\usage{
.handle_http_response(response)
}
\arguments{
\item{response}{A response object from \code{crul::HttpClient$get()}.}
}
\value{
Invisibly returns the response if status code is < 400,
otherwise throws an error.
}
\description{
Validates HTTP response status codes and extracts error messages.
}
\keyword{internal}
