% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.is_boolean}
\alias{.is_boolean}
\title{Boolean}
\usage{
.is_boolean(x)
}
\arguments{
\item{x}{an object to check.}
}
\value{
a logical value indicating whether provided object is a \code{Boolean}.
}
\description{
Checks if provided object is a Boolean i.e. a length-one logical vector.
}
\note{
Taken from
\url{https://github.com/Rapporter/rapportools/blob/master/R/utils.R}
}
\examples{
is_boolean(TRUE) # [1] TRUE
# the following will work on most systems, unless you have tweaked global Rprofile
.is_boolean(T) # [1] TRUE
.is_boolean(1) # [1] FALSE
}
\keyword{internal}
