% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.send_mgmt_query}
\alias{.send_mgmt_query}
\title{Send Query to POWER Management API}
\usage{
.send_mgmt_query(.url)
}
\arguments{
\item{.url}{A character string of the URL to be used for the API query.}
}
\value{
The HTTP response object from the POWER server containing either
an error message or the requested data.
}
\description{
Sends the query to the POWER management API.
}
\details{
Note: The management API endpoints do not accept parameters.
}
\keyword{internal}
