% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ag_data.R
\name{get_ag_data}
\alias{get_ag_data}
\title{Get Data Suitable for Use in Agriculture Simulation Modelling from the POWER Web API}
\usage{
get_ag_data(lonlat, dates)
}
\arguments{
\item{lonlat}{A numeric vector of geographic coordinates for a cell or region
entered as x, y coordinates.  See argument details for more.}

\item{dates}{A character vector of start and end dates in that order,\cr
\emph{e.g.}, \code{dates = c("1983-01-01", "2017-12-31")}.  See argument
details for more.}
}
\value{
A data frame as a code{\link[tibble]{tibble}} of data suitable for
  creating a .met or .icasa file for use in \acronym{APSIM} or
  \acronym{DSSAT} crop modelling software.
}
\description{
Get \acronym{POWER} values for a single point or region and create
  an object suitable for the creation of an \acronym{APSIM-X} \code{met} file
  or a \acronym{DSSAT} \acronym{ICASA} object for use in crop modelling.
  Values are returned as a tidy data frame \code{\link[tibble]{tibble}}
  object.
}
\details{
This function is a shortcut for \code{\link{get_power}} that
  simplifies the querying of the \acronym{POWER} \acronym{API}, fetching
  the most commonly used data for agricultural modelling without the user
  needing to specify every parameter.

  The weather values from \acronym{POWER} for temperature are 2 metre max and
  min temperatures, \dQuote{T2M_MAX} and \dQuote{T2M_MIN}; radiation,
  \dQuote{ALLSKY_SFC_SW_DWN}; rain, \dQuote{PRECTOT}; relative humidity at 2
  metres, \dQuote{RH2M}; and wind at 2 metres \dQuote{WS2M} from the
  \acronym{POWER} \sQuote{AG} community on a daily time-step.

  Users can use functions from \pkg{apsimx} or \pkg{DSSAT} to generate input
  files useful in crop simulation modelling software like \acronym{APSIM-X}
  or \acronym{DSSAT}.
}
\section{Argument details for \code{lonlat}}{

\describe{
  \item{For a single point}{To get a specific cell, 1/2 x 1/2 degree, supply
  a length-two numeric vector giving the decimal degree longitude and
  latitude in that order for data to download,\cr
  \emph{e.g.}, \code{lonlat = c(151.81, -27.48)}.}

  \item{For regional coverage}{To get a region, supply a length-four numeric
  vector as lower left (lon, lat) and upper right (lon, lat) coordinates,
  \emph{e.g.}, \code{lonlat = c(xmin, ymin, xmax, ymax)} in that order for a
  given region, \emph{e.g.}, a bounding box for the south western corner of
  Australia: \code{lonlat = c(112.5, -55.5, 115.5, -50.5)}. \emph{Max
  bounding box is 10 x 10 degrees} of 1/2 x 1/2 degree data, \emph{i.e.}, 100
  points maximum in total.}
}
}

\section{Argument details for \code{dates}}{
 If \code{dates} is unspecified,
  defaults to a start date of 1983-01-01 (the earliest available data) and an
  end date of current date according to the system.

  If one date only is provided, it will be treated as both the start date and
  the end date and only a single day's values will be returned.
}

\examples{
# Get data to create a .met or .icasa format file for Kingsthorpe, Qld
# from 1985-01-01 to 1985-06-30

\donttest{
get_ag_data(lonlat = c(151.81, -27.48),
           dates = c("1985-01-01", "1985-12-31")
           )
}

}
\author{
Sparks, A. H. \email{adamhsparks@gmail.com}
}
