% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icasa.R
\name{get_icasa}
\alias{get_icasa}
\title{Get Data Suitable for Creating a DSSAT ICASA File from the POWER Web API}
\usage{
get_icasa(lonlat, dates)
}
\arguments{
\item{lonlat}{A numeric vector of geographic coordinates for a cell or region
entered as x, y coordinates.  See argument details for more.}

\item{dates}{A character vector of start and end dates in that order,\cr
\emph{e.g.}, \code{dates = c("1983-01-01", "2017-12-31")}.  See argument
details for  more.}
}
\value{
An \code{data frame} object of data suitable for use in
 \acronym{DSSAT} crop modelling.
}
\description{
Get \acronym{POWER} values for a single point or region and create an
  \acronym{ICASA} format text file suitable for use in \acronym{DSSAT} for
  crop modelling.
}
\details{
This function is essentially a wrapper for \code{\link{get_power}}
  that queries the \acronym{POWER} \acronym{API} and writes a \acronym{DSSAT}
  \acronym{ICASA} weather file to disk.  All necessary \code{pars} are
  automatically included in the query.

  Further details for each of the arguments are provided in their
  respective sections following below.
}
\section{Argument details for \code{lonlat}}{

\describe{
  \item{For a single point}{To get a specific cell, 1/2 x 1/2 degree, supply
  a length-two numeric vector giving the decimal degree longitude and
  latitude in that order for data to download,\cr
  \emph{e.g.}, \code{lonlat = c(151.81, -27.48)}.}

  \item{For regional coverage}{To get a region, supply a length-four numeric
  vector as lower left (lon, lat) and upper right (lon, lat) coordinates,
  \emph{e.g.}, \code{lonlat = c(xmin, ymin, xmax, ymax)} in that order for a
  given region, \emph{e.g.}, a bounding box for the southwestern corner of
  Australia: \code{lonlat = c(112.5, -55.5, 115.5, -50.5)}. \emph{Max
  bounding box is 10 x 10 degrees} of 1/2 x 1/2 degree data, \emph{i.e.}, 100
  points maximum in total.}
}
}

\section{Argument details for \code{dates}}{
 If \code{dates} is unspecified,
  defaults to a start date of 1983-01-01 (the earliest available data) and an
  end date of current date according to the system. If one date only is
  provided, it will be treated as both the start date and the end date and
  only a single day's values will be returned.
}

\examples{
# Get data for an ICASA file for Kingsthorpe,
# Qld from 1985-01-01 to 1985-06-30
# and save it in the current R session
# tempdir() as ICASA_example.txt

\donttest{
create_icasa(lonlat = c(151.81, -27.48),
            dates = c("1985-01-01", "1985-12-31"),
            dsn = tempdir(),
            file_out = "ICASA_example.txt"
            )
}

}
\seealso{
\code{\link{get_met}} Get Data Suitable for Creating an
 \acronym{APSIM} .met file from the \acronym{POWER} Web \acronym{API}.
}
\author{
Sparks, A. H. \email{adamhsparks@gmail.com}
}
