% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Adds a \%notin\% Function}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{A character string to match.}

\item{table}{A table containing values to match \code{x} against.}
}
\value{
A logical vector, indicating if a mismatch was located for any
element of x: thus the values are \code{TRUE} or \code{FALSE} and never \code{NA}.
}
\description{
Negates \code{\%in\%} for easier (miss)matching.
}
\keyword{internal}
