% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasapower-defunct.R
\name{nasapower-defunct}
\alias{nasapower-defunct}
\alias{create_met}
\alias{create_icasa}
\title{Defunct functions in \CRANpkg{nasapower}

These functions are gone, no longer available.

\itemize{
 \item \code{\link{create_icasa}}: This function is defunct.
 \item \code{\link{create_met}}: This function is defunct.
}}
\usage{
create_met(...)

create_icasa(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct functions in \CRANpkg{nasapower}

These functions are gone, no longer available.

\itemize{
 \item \code{\link{create_icasa}}: This function is defunct.
 \item \code{\link{create_met}}: This function is defunct.
}
}
\details{
`create_met()` is now defunct and has been removed from
 \CRANpkg{nasapower}

`create_icasa()` is now defunct and has been removed from
 \CRANpkg{nasapower}
}
