% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\docType{package}
\name{nasapower-deprecated}
\alias{nasapower-deprecated}
\alias{get_cell}
\alias{nasapower-deprecated-package}
\alias{nasapower-deprecated}
\alias{get_region}
\title{Deprecated function(s) in the nasapower package}
\usage{
get_cell()

get_region()
}
\description{
These functions are now deprecated in \pkg{nasapower}.
}
\section{Details}{

\tabular{rl}{
  \code{get_cell} \tab now superceded by \code{get_power}\cr
  \code{get_region} \tab now superceded by \code{get_power}\cr
}
}

