% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasapower-package.R
\docType{package}
\name{nasapower-package}
\alias{nasapower}
\alias{nasapower-package}
\title{nasapower: NASA POWER API Client}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Client for 'NASA' 'POWER' global meteorology, surface solar
    energy and climatology data 'API'.  'POWER' (Prediction Of Worldwide Energy
    Resource) data are freely available global meteorology and surface solar
    energy climatology data for download with a resolution of 1/2 by 1/2 arc
    degree longitude and latitude and are funded through the 'NASA' Earth
    Science Directorate Applied Science Program.  For more on the data
    themselves, a web-based data viewer and web access, please see
    <https://power.larc.nasa.gov/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/nasapower/}
  \item Report bugs at \url{https://github.com/ropensci/nasapower/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Other contributors:
\itemize{
  \item Scott Chamberlain \email{myrmecocystus@gmail.com} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID}) [reviewer]
  \item Hazel Kavili [reviewer]
  \item Alison Boyer [reviewer]
}

}
\keyword{internal}
