% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasapower-package.R
\docType{package}
\name{nasapower}
\alias{nasapower}
\alias{nasapower-package}
\title{Downloads NASA-POWER Agroclimatology Variables and Returns a Tidy Data Frame}
\description{
Downloads NASA-POWER Agroclimatology Variables and Returns a Tidy Data Frame
}
\seealso{
\itemize{
\item \link{get_cell} Download NASA-POWER Agroclimatology Variables for a
Given Cell and Return a Tidy Data Frame
\item \link{get_region} Download NASA-POWER Agroclimatology Variables for a
Given Region and Return a Tidy Data Frame
}
}
\author{
Adam H Sparks
}
