% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{NASA POWER parameters available for download}
\format{
A list with 389 weather and climate parameters contained within the
database.
}
\source{
\url{https://power.larc.nasa.gov/beta/api/system/manager/system/parameters}
}
\usage{
parameters
}
\description{
An \R \code{list} object of \acronym{POWER} parameters and metadata available for
querying from the \acronym{POWER} database.
}
\section{Valid parameters}{

\describe{
\item{
AOD_55_ADJ
}{
Adjusted Aerosol Optical Depth 55
}
\item{
AOD_55
}{
Aerosol Optical Depth 55
}
\item{
AOD_84
}{
Aerosol Optical Depth 84
}
\item{
AIRMASS
}{
Air Mass
}
\item{
ALLSKY_KT
}{
All Sky Insolation Clearness Index
}
\item{
ALLSKY_KT_MAX
}{
All Sky Insolation Clearness Index Maximum
}
\item{
ALLSKY_KT_MIN
}{
All Sky Insolation Clearness Index Minimum
}
\item{
ALLSKY_KT_SD
}{
All Sky Insolation Clearness Index Standard Deviation
}
\item{
ALLSKY_NKT
}{
All Sky Normalized Insolation Clearness Index
}
\item{
ALLSKY_SRF_ALB
}{
All Sky Surface Albedo
}
\item{
ALLSKY_SFC_LW_DWN
}{
All Sky Surface Longwave Downward Irradiance
}
\item{
ALLSKY_SFC_LW_DWN_MAX
}{
All Sky Surface Longwave Downward Irradiance Maximum
}
\item{
ALLSKY_SFC_LW_DWN_MIN
}{
All Sky Surface Longwave Downward Irradiance Minimum
}
\item{
ALLSKY_SFC_LW_DWN_SD
}{
All Sky Surface Longwave Downward Irradiance Standard Deviation
}
\item{
ALLSKY_SFC_LW_UP
}{
All Sky Surface Longwave Upward Irradiance
}
\item{
ALLSKY_SFC_LW_UP_MAX
}{
All Sky Surface Longwave Upward Irradiance Maximum
}
\item{
ALLSKY_SFC_LW_UP_MIN
}{
All Sky Surface Longwave Upward Irradiance Minimum
}
\item{
ALLSKY_SFC_LW_UP_SD
}{
All Sky Surface Longwave Upward Irradiance Standard Deviation
}
\item{
ALLSKY_SFC_PAR_TOT
}{
All Sky Surface PAR Total
}
\item{
ALLSKY_SFC_PAR_TOT_MAX
}{
All Sky Surface PAR Total Maximum
}
\item{
ALLSKY_SFC_PAR_TOT_MIN
}{
All Sky Surface PAR Total Minimum
}
\item{
ALLSKY_SFC_PAR_TOT_SD
}{
All Sky Surface PAR Total Standard Deviation
}
\item{
ALLSKY_SFC_SW_DIFF
}{
All Sky Surface Shortwave Diffuse Irradiance
}
\item{
ALLSKY_SFC_SW_DIFF_MAX
}{
All Sky Surface Shortwave Diffuse Irradiance Maximum
}
\item{
ALLSKY_SFC_SW_DIFF_MIN
}{
All Sky Surface Shortwave Diffuse Irradiance Minimum
}
\item{
ALLSKY_SFC_SW_DIFF_SD
}{
All Sky Surface Shortwave Diffuse Irradiance Standard Deviation
}
\item{
ALLSKY_SFC_SW_DIRH
}{
All Sky Surface Shortwave Direct Horizontal Irradiance
}
\item{
ALLSKY_SFC_SW_DIRH_MAX
}{
All Sky Surface Shortwave Direct Horizontal Irradiance Maximum
}
\item{
ALLSKY_SFC_SW_DIRH_MIN
}{
All Sky Surface Shortwave Direct Horizontal Irradiance Minimum
}
\item{
ALLSKY_SFC_SW_DIRH_SD
}{
All Sky Surface Shortwave Direct Horizontal Irradiance Standard Deviation
}
\item{
ALLSKY_SFC_SW_DNI
}{
All Sky Surface Shortwave Downward Direct Normal Irradiance
}
\item{
ALLSKY_SFC_SW_DNI_MAX
}{
All Sky Surface Shortwave Downward Direct Normal Irradiance Maximum
}
\item{
ALLSKY_SFC_SW_DNI_MAX_RD
}{
All Sky Surface Shortwave Downward Direct Normal Irradiance Maximum Relative
Difference
}
\item{
ALLSKY_SFC_SW_DNI_MIN
}{
All Sky Surface Shortwave Downward Direct Normal Irradiance Minimum
}
\item{
ALLSKY_SFC_SW_DNI_MIN_RD
}{
All Sky Surface Shortwave Downward Direct Normal Irradiance Minimum Relative
Difference
}
\item{
ALLSKY_SFC_SW_DNI_SD
}{
All Sky Surface Shortwave Downward Direct Normal Irradiance Standard
Deviation
}
\item{
ALLSKY_SFC_SW_DWN
}{
All Sky Surface Shortwave Downward Irradiance
}
\item{
ALLSKY_SFC_SW_DWN_MAX
}{
All Sky Surface Shortwave Downward Irradiance Maximum
}
\item{
ALLSKY_SFC_SW_DWN_MAX_RD
}{
All Sky Surface Shortwave Downward Irradiance Maximum Relative Difference
}
\item{
ALLSKY_SFC_SW_DWN_MIN
}{
All Sky Surface Shortwave Downward Irradiance Minimum
}
\item{
ALLSKY_SFC_SW_DWN_MIN_RD
}{
All Sky Surface Shortwave Downward Irradiance Minimum Relative Difference
}
\item{
ALLSKY_SFC_SW_DWN_SD
}{
All Sky Surface Shortwave Downward Irradiance Standard Deviation
}
\item{
ALLSKY_SFC_SW_DWN_00
}{
All Sky Surface Shortwave Downward Irradiance at 00 GMT
}
\item{
ALLSKY_SFC_SW_DWN_01
}{
All Sky Surface Shortwave Downward Irradiance at 01 GMT
}
\item{
ALLSKY_SFC_SW_DWN_02
}{
All Sky Surface Shortwave Downward Irradiance at 02 GMT
}
\item{
ALLSKY_SFC_SW_DWN_03
}{
All Sky Surface Shortwave Downward Irradiance at 03 GMT
}
\item{
ALLSKY_SFC_SW_DWN_04
}{
All Sky Surface Shortwave Downward Irradiance at 04 GMT
}
\item{
ALLSKY_SFC_SW_DWN_05
}{
All Sky Surface Shortwave Downward Irradiance at 05 GMT
}
\item{
ALLSKY_SFC_SW_DWN_06
}{
All Sky Surface Shortwave Downward Irradiance at 06 GMT
}
\item{
ALLSKY_SFC_SW_DWN_07
}{
All Sky Surface Shortwave Downward Irradiance at 07 GMT
}
\item{
ALLSKY_SFC_SW_DWN_08
}{
All Sky Surface Shortwave Downward Irradiance at 08 GMT
}
\item{
ALLSKY_SFC_SW_DWN_09
}{
All Sky Surface Shortwave Downward Irradiance at 09 GMT
}
\item{
ALLSKY_SFC_SW_DWN_10
}{
All Sky Surface Shortwave Downward Irradiance at 10 GMT
}
\item{
ALLSKY_SFC_SW_DWN_11
}{
All Sky Surface Shortwave Downward Irradiance at 11 GMT
}
\item{
ALLSKY_SFC_SW_DWN_12
}{
All Sky Surface Shortwave Downward Irradiance at 12 GMT
}
\item{
ALLSKY_SFC_SW_DWN_13
}{
All Sky Surface Shortwave Downward Irradiance at 13 GMT
}
\item{
ALLSKY_SFC_SW_DWN_14
}{
All Sky Surface Shortwave Downward Irradiance at 14 GMT
}
\item{
ALLSKY_SFC_SW_DWN_15
}{
All Sky Surface Shortwave Downward Irradiance at 15 GMT
}
\item{
ALLSKY_SFC_SW_DWN_16
}{
All Sky Surface Shortwave Downward Irradiance at 16 GMT
}
\item{
ALLSKY_SFC_SW_DWN_17
}{
All Sky Surface Shortwave Downward Irradiance at 17 GMT
}
\item{
ALLSKY_SFC_SW_DWN_18
}{
All Sky Surface Shortwave Downward Irradiance at 18 GMT
}
\item{
ALLSKY_SFC_SW_DWN_19
}{
All Sky Surface Shortwave Downward Irradiance at 19 GMT
}
\item{
ALLSKY_SFC_SW_DWN_20
}{
All Sky Surface Shortwave Downward Irradiance at 20 GMT
}
\item{
ALLSKY_SFC_SW_DWN_21
}{
All Sky Surface Shortwave Downward Irradiance at 21 GMT
}
\item{
ALLSKY_SFC_SW_DWN_22
}{
All Sky Surface Shortwave Downward Irradiance at 22 GMT
}
\item{
ALLSKY_SFC_SW_DWN_23
}{
All Sky Surface Shortwave Downward Irradiance at 23 GMT
}
\item{
ALLSKY_SFC_SW_DWN_HR
}{
All Sky Surface Shortwave Downward Irradiance at GMT Times
}
\item{
ALLSKY_SFC_SW_UP
}{
All Sky Surface Shortwave Upward Irradiance
}
\item{
ALLSKY_SFC_SW_UP_MAX
}{
All Sky Surface Shortwave Upward Irradiance Maximum
}
\item{
ALLSKY_SFC_SW_UP_MIN
}{
All Sky Surface Shortwave Upward Irradiance Minimum
}
\item{
ALLSKY_SFC_SW_UP_SD
}{
All Sky Surface Shortwave Upward Irradiance Standard Deviation
}
\item{
ALLSKY_SFC_UV_INDEX
}{
All Sky Surface UV Index
}
\item{
ALLSKY_SFC_UV_INDEX_MAX
}{
All Sky Surface UV Index Maximum
}
\item{
ALLSKY_SFC_UV_INDEX_MIN
}{
All Sky Surface UV Index Minimum
}
\item{
ALLSKY_SFC_UV_INDEX_SD
}{
All Sky Surface UV Index Standard Deviation
}
\item{
ALLSKY_SFC_UVA
}{
All Sky Surface UVA Irradiance
}
\item{
ALLSKY_SFC_UVA_MAX
}{
All Sky Surface UVA Irradiance Maximum
}
\item{
ALLSKY_SFC_UVA_MIN
}{
All Sky Surface UVA Irradiance Minimum
}
\item{
ALLSKY_SFC_UVA_SD
}{
All Sky Surface UVA Irradiance Standard Deviation
}
\item{
ALLSKY_SFC_UVB
}{
All Sky Surface UVB Irradiance
}
\item{
ALLSKY_SFC_UVB_MAX
}{
All Sky Surface UVB Irradiance Maximum
}
\item{
ALLSKY_SFC_UVB_MIN
}{
All Sky Surface UVB Irradiance Minimum
}
\item{
ALLSKY_SFC_UVB_SD
}{
All Sky Surface UVB Irradiance Standard Deviation
}
\item{
SG_MID_COZ_ZEN_ANG
}{
Average Cosine Solar Zenith Angle At Mid - Time Between Sunrise And Solar
Noon for Climatological Month
}
\item{
SG_DEC
}{
Average Declination for Climatological Month
}
\item{
SG_HRZ_HR
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological Month
}
\item{
SG_HRZ_00
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 00 GMT
}
\item{
SG_HRZ_01
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 01 GMT
}
\item{
SG_HRZ_02
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 02 GMT
}
\item{
SG_HRZ_03
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 03 GMT
}
\item{
SG_HRZ_04
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 04 GMT
}
\item{
SG_HRZ_05
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 05 GMT
}
\item{
SG_HRZ_06
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 06 GMT
}
\item{
SG_HRZ_07
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 07 GMT
}
\item{
SG_HRZ_08
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 08 GMT
}
\item{
SG_HRZ_09
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 09 GMT
}
\item{
SG_HRZ_10
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 10 GMT
}
\item{
SG_HRZ_11
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 11 GMT
}
\item{
SG_HRZ_12
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 12 GMT
}
\item{
SG_HRZ_13
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 13 GMT
}
\item{
SG_HRZ_14
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 14 GMT
}
\item{
SG_HRZ_15
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 15 GMT
}
\item{
SG_HRZ_16
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 16 GMT
}
\item{
SG_HRZ_17
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 17 GMT
}
\item{
SG_HRZ_18
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 18 GMT
}
\item{
SG_HRZ_19
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 19 GMT
}
\item{
SG_HRZ_20
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 20 GMT
}
\item{
SG_HRZ_21
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 21 GMT
}
\item{
SG_HRZ_22
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 22 GMT
}
\item{
SG_HRZ_23
}{
Average Hourly Solar Angles Relative To The Horizon for Climatological
Month at 23 GMT
}
\item{
SG_SAA_HR
}{
Average Solar Azimuth Angle for Climatological Month
}
\item{
SG_SAA_00
}{
Average Solar Azimuth Angle for Climatological Month at 00 GMT
}
\item{
SG_SAA_01
}{
Average Solar Azimuth Angle for Climatological Month at 01 GMT
}
\item{
SG_SAA_02
}{
Average Solar Azimuth Angle for Climatological Month at 02 GMT
}
\item{
SG_SAA_03
}{
Average Solar Azimuth Angle for Climatological Month at 03 GMT
}
\item{
SG_SAA_04
}{
Average Solar Azimuth Angle for Climatological Month at 04 GMT
}
\item{
SG_SAA_05
}{
Average Solar Azimuth Angle for Climatological Month at 05 GMT
}
\item{
SG_SAA_06
}{
Average Solar Azimuth Angle for Climatological Month at 06 GMT
}
\item{
SG_SAA_07
}{
Average Solar Azimuth Angle for Climatological Month at 07 GMT
}
\item{
SG_SAA_08
}{
Average Solar Azimuth Angle for Climatological Month at 08 GMT
}
\item{
SG_SAA_09
}{
Average Solar Azimuth Angle for Climatological Month at 09 GMT
}
\item{
SG_SAA_10
}{
Average Solar Azimuth Angle for Climatological Month at 10 GMT
}
\item{
SG_SAA_11
}{
Average Solar Azimuth Angle for Climatological Month at 11 GMT
}
\item{
SG_SAA_12
}{
Average Solar Azimuth Angle for Climatological Month at 12 GMT
}
\item{
SG_SAA_13
}{
Average Solar Azimuth Angle for Climatological Month at 13 GMT
}
\item{
SG_SAA_14
}{
Average Solar Azimuth Angle for Climatological Month at 14 GMT
}
\item{
SG_SAA_15
}{
Average Solar Azimuth Angle for Climatological Month at 15 GMT
}
\item{
SG_SAA_16
}{
Average Solar Azimuth Angle for Climatological Month at 16 GMT
}
\item{
SG_SAA_17
}{
Average Solar Azimuth Angle for Climatological Month at 17 GMT
}
\item{
SG_SAA_18
}{
Average Solar Azimuth Angle for Climatological Month at 18 GMT
}
\item{
SG_SAA_19
}{
Average Solar Azimuth Angle for Climatological Month at 19 GMT
}
\item{
SG_SAA_20
}{
Average Solar Azimuth Angle for Climatological Month at 20 GMT
}
\item{
SG_SAA_21
}{
Average Solar Azimuth Angle for Climatological Month at 21 GMT
}
\item{
SG_SAA_22
}{
Average Solar Azimuth Angle for Climatological Month at 22 GMT
}
\item{
SG_SAA_23
}{
Average Solar Azimuth Angle for Climatological Month at 23 GMT
}
\item{
SG_HR_SET_ANG
}{
Average Sunset Hour Angle for Climatological Month
}
\item{
CLRSKY_DAYS
}{
Clear Sky Day
}
\item{
CLRSKY_KT
}{
Clear Sky Insolation Clearness Index
}
\item{
CLRSKY_KT_MAX
}{
Clear Sky Insolation Clearness Index Maximum
}
\item{
CLRSKY_KT_MIN
}{
Clear Sky Insolation Clearness Index Minimum
}
\item{
CLRSKY_KT_SD
}{
Clear Sky Insolation Clearness Index Standard Deviation
}
\item{
CLRSKY_NKT
}{
Clear Sky Normalized Insolation Clearness Index
}
\item{
CLRSKY_SRF_ALB
}{
Clear Sky Surface Albedo
}
\item{
CLRSKY_SFC_LW_DWN
}{
Clear Sky Surface Longwave Downward Irradiance
}
\item{
CLRSKY_SFC_LW_DWN_MAX
}{
Clear Sky Surface Longwave Downward Irradiance Maximum
}
\item{
CLRSKY_SFC_LW_DWN_MIN
}{
Clear Sky Surface Longwave Downward Irradiance Minimum
}
\item{
CLRSKY_SFC_LW_DWN_SD
}{
Clear Sky Surface Longwave Downward Irradiance Standard Deviation
}
\item{
CLRSKY_SFC_LW_UP
}{
Clear Sky Surface Longwave Upward Irradiance
}
\item{
CLRSKY_SFC_LW_UP_MAX
}{
Clear Sky Surface Longwave Upward Irradiance Maximum
}
\item{
CLRSKY_SFC_LW_UP_MIN
}{
Clear Sky Surface Longwave Upward Irradiance Minimum
}
\item{
CLRSKY_SFC_LW_UP_SD
}{
Clear Sky Surface Longwave Upward Irradiance Standard Deviation
}
\item{
CLRSKY_SFC_PAR_TOT
}{
Clear Sky Surface PAR Total
}
\item{
CLRSKY_SFC_PAR_TOT_MAX
}{
Clear Sky Surface PAR Total Maximum
}
\item{
CLRSKY_SFC_PAR_TOT_MIN
}{
Clear Sky Surface PAR Total Minimum
}
\item{
CLRSKY_SFC_PAR_TOT_SD
}{
Clear Sky Surface PAR Total Standard Deviation
}
\item{
CLRSKY_SFC_SW_DWN
}{
Clear Sky Surface Shortwave Downward Irradiance
}
\item{
CLRSKY_SFC_SW_DWN_MAX
}{
Clear Sky Surface Shortwave Downward Irradiance Maximum
}
\item{
CLRSKY_SFC_SW_DWN_MIN
}{
Clear Sky Surface Shortwave Downward Irradiance Minimum
}
\item{
CLRSKY_SFC_SW_DWN_SD
}{
Clear Sky Surface Shortwave Downward Irradiance Standard Deviation
}
\item{
CLRSKY_SFC_SW_UP
}{
Clear Sky Surface Shortwave Upward Irradiance
}
\item{
CLRSKY_SFC_SW_UP_MAX
}{
Clear Sky Surface Shortwave Upward Irradiance Maximum
}
\item{
CLRSKY_SFC_SW_UP_MIN
}{
Clear Sky Surface Shortwave Upward Irradiance Minimum
}
\item{
CLRSKY_SFC_SW_UP_SD
}{
Clear Sky Surface Shortwave Upward Irradiance Standard Deviation
}
\item{
CLOUD_AMT
}{
Cloud Amount
}
\item{
CLOUD_AMT_00
}{
Cloud Amount at 00 GMT
}
\item{
CLOUD_AMT_01
}{
Cloud Amount at 01 GMT
}
\item{
CLOUD_AMT_02
}{
Cloud Amount at 02 GMT
}
\item{
CLOUD_AMT_03
}{
Cloud Amount at 03 GMT
}
\item{
CLOUD_AMT_04
}{
Cloud Amount at 04 GMT
}
\item{
CLOUD_AMT_05
}{
Cloud Amount at 05 GMT
}
\item{
CLOUD_AMT_06
}{
Cloud Amount at 06 GMT
}
\item{
CLOUD_AMT_07
}{
Cloud Amount at 07 GMT
}
\item{
CLOUD_AMT_08
}{
Cloud Amount at 08 GMT
}
\item{
CLOUD_AMT_09
}{
Cloud Amount at 09 GMT
}
\item{
CLOUD_AMT_10
}{
Cloud Amount at 10 GMT
}
\item{
CLOUD_AMT_11
}{
Cloud Amount at 11 GMT
}
\item{
CLOUD_AMT_12
}{
Cloud Amount at 12 GMT
}
\item{
CLOUD_AMT_13
}{
Cloud Amount at 13 GMT
}
\item{
CLOUD_AMT_14
}{
Cloud Amount at 14 GMT
}
\item{
CLOUD_AMT_15
}{
Cloud Amount at 15 GMT
}
\item{
CLOUD_AMT_16
}{
Cloud Amount at 16 GMT
}
\item{
CLOUD_AMT_17
}{
Cloud Amount at 17 GMT
}
\item{
CLOUD_AMT_18
}{
Cloud Amount at 18 GMT
}
\item{
CLOUD_AMT_19
}{
Cloud Amount at 19 GMT
}
\item{
CLOUD_AMT_20
}{
Cloud Amount at 20 GMT
}
\item{
CLOUD_AMT_21
}{
Cloud Amount at 21 GMT
}
\item{
CLOUD_AMT_22
}{
Cloud Amount at 22 GMT
}
\item{
CLOUD_AMT_23
}{
Cloud Amount at 23 GMT
}
\item{
CLOUD_AMT_DAY
}{
Cloud Amount at Daytime
}
\item{
CLOUD_AMT_HR
}{
Cloud Amount at GMT Times
}
\item{
CLOUD_AMT_NIGHT
}{
Cloud Amount at Nighttime
}
\item{
CLOUD_OD
}{
Cloud Optical Visible Depth
}
\item{
CLOUD_OD_00
}{
Cloud Optical Visible Depth at 00 GMT
}
\item{
CLOUD_OD_01
}{
Cloud Optical Visible Depth at 01 GMT
}
\item{
CLOUD_OD_02
}{
Cloud Optical Visible Depth at 02 GMT
}
\item{
CLOUD_OD_03
}{
Cloud Optical Visible Depth at 03 GMT
}
\item{
CLOUD_OD_04
}{
Cloud Optical Visible Depth at 04 GMT
}
\item{
CLOUD_OD_05
}{
Cloud Optical Visible Depth at 05 GMT
}
\item{
CLOUD_OD_06
}{
Cloud Optical Visible Depth at 06 GMT
}
\item{
CLOUD_OD_07
}{
Cloud Optical Visible Depth at 07 GMT
}
\item{
CLOUD_OD_08
}{
Cloud Optical Visible Depth at 08 GMT
}
\item{
CLOUD_OD_09
}{
Cloud Optical Visible Depth at 09 GMT
}
\item{
CLOUD_OD_10
}{
Cloud Optical Visible Depth at 10 GMT
}
\item{
CLOUD_OD_11
}{
Cloud Optical Visible Depth at 11 GMT
}
\item{
CLOUD_OD_12
}{
Cloud Optical Visible Depth at 12 GMT
}
\item{
CLOUD_OD_13
}{
Cloud Optical Visible Depth at 13 GMT
}
\item{
CLOUD_OD_14
}{
Cloud Optical Visible Depth at 14 GMT
}
\item{
CLOUD_OD_15
}{
Cloud Optical Visible Depth at 15 GMT
}
\item{
CLOUD_OD_16
}{
Cloud Optical Visible Depth at 16 GMT
}
\item{
CLOUD_OD_17
}{
Cloud Optical Visible Depth at 17 GMT
}
\item{
CLOUD_OD_18
}{
Cloud Optical Visible Depth at 18 GMT
}
\item{
CLOUD_OD_19
}{
Cloud Optical Visible Depth at 19 GMT
}
\item{
CLOUD_OD_20
}{
Cloud Optical Visible Depth at 20 GMT
}
\item{
CLOUD_OD_21
}{
Cloud Optical Visible Depth at 21 GMT
}
\item{
CLOUD_OD_22
}{
Cloud Optical Visible Depth at 22 GMT
}
\item{
CLOUD_OD_23
}{
Cloud Optical Visible Depth at 23 GMT
}
\item{
CLOUD_OD_HR
}{
Cloud Optical Visible Depth at GMT Times
}
\item{
CDD0
}{
Cooling Degree Days Above 0 C
}
\item{
CDD10
}{
Cooling Degree Days Above 10 C
}
\item{
CDD18_3
}{
Cooling Degree Days Above 18.3 C
}
\item{
PSC
}{
Corrected Atmospheric Pressure (Adjusted For Site Elevation)
}
\item{
WSC
}{
Corrected Wind Speed (Adjusted For Elevation)
}
\item{
SG_DAY_COZ_ZEN_AVG
}{
Daylight Average Of Hourly Cosine Solar Zenith Angles for Climatological
Month
}
\item{
T2MDEW
}{
Dew / Frost Point at 2 Meters
}
\item{
DIFFUSE_ILLUMINANCE
}{
Diffuse Illuminance
}
\item{
DIRECT_ILLUMINANCE
}{
Direct Illuminance
}
\item{
TS
}{
Earth Skin Temperature
}
\item{
TS_ADJ
}{
Earth Skin Temperature Adjusted
}
\item{
TS_AMP
}{
Earth Skin Temperature Amplitude
}
\item{
TS_MAX
}{
Earth Skin Temperature Maximum
}
\item{
TS_MAX_AVG
}{
Earth Skin Temperature Maximum Average
}
\item{
TS_MIN
}{
Earth Skin Temperature Minimum
}
\item{
TS_MIN_AVG
}{
Earth Skin Temperature Minimum Average
}
\item{
TS_RANGE
}{
Earth Skin Temperature Range
}
\item{
TS_RANGE_AVG
}{
Earth Skin Temperature Range Average
}
\item{
U10M
}{
Eastward Wind at 10 Meters
}
\item{
U2M
}{
Eastward Wind at 2 Meters
}
\item{
U50M
}{
Eastward Wind at 50 Meters
}
\item{
NO_SUN_BLACKDAYS_MAX
}{
Equivalent No - Sun Days
}
\item{
EQUIV_NO_SUN_CONSEC_01
}{
Equivalent No - Sun Days Over A Consecutive 1 - day Period
}
\item{
EQUIV_NO_SUN_CONSEC_14
}{
Equivalent No - Sun Days Over A Consecutive 14 - day Period
}
\item{
EQUIV_NO_SUN_CONSEC_21
}{
Equivalent No - Sun Days Over A Consecutive 21 - day Period
}
\item{
EQUIV_NO_SUN_CONSEC_03
}{
Equivalent No - Sun Days Over A Consecutive 3 - day Period
}
\item{
EQUIV_NO_SUN_CONSEC_07
}{
Equivalent No - Sun Days Over A Consecutive 7 - day Period
}
\item{
EQUIV_NO_SUN_CONSEC_MONTH
}{
Equivalent No - Sun Days Over A Consecutive Month Period
}
\item{
EVLAND
}{
Evaporation Land
}
\item{
EVPTRNS
}{
Evapotranspiration Energy Flux
}
\item{
FROST_DAYS
}{
Frost Days
}
\item{
GLOBAL_ILLUMINANCE
}{
Global Illuminance
}
\item{
HDD0
}{
Heating Degree Days Below 0 C
}
\item{
HDD10
}{
Heating Degree Days Below 10 C
}
\item{
HDD18_3
}{
Heating Degree Days Below 18.3 C
}
\item{
FRSEAICE
}{
Ice Covered Fraction
}
\item{
FRSNO
}{
Land Snowcover Fraction
}
\item{
MAX_EQUIV_NO_SUN_DEFICIT
}{
Maximum Equivalent No - Sun Days Deficit
}
\item{
SG_HRZ_MAX
}{
Maximum Solar Angle Relative To The Horizon for Climatological Month
}
\item{
SI_EF_MAX_TILTED_SURFACE_TRACKER
}{
Maximum Solar Irradiance Irradiance Tracking the Sun
}
\item{
SI_EF_MAX_TILTED_SURFACE_OPTIMAL
}{
Maximum Solar Irradiance Optimal
}
\item{
SI_EF_MAX_TILTED_SURFACE_OPTIMAL_ANG
}{
Maximum Solar Irradiance Optimal Angle
}
\item{
SI_EF_MAX_TILTED_SURFACE_OPTIMAL_ANG_ORT
}{
Maximum Solar Irradiance Tilted Surface Orientation
}
\item{
SI_EF_MAX_TILTED_SURFACE_HORIZONTAL
}{
Maximum Solar Irradiance for Equator Facing Horizontal Surface
}
\item{
SI_EF_MAX_TILTED_SURFACE_LAT_MINUS15
}{
Maximum Solar Irradiance for Equator Facing Latitude Minus 15 Tilt
}
\item{
SI_EF_MAX_TILTED_SURFACE_LAT_PLUS15
}{
Maximum Solar Irradiance for Equator Facing Latitude Plus 15 Tilt
}
\item{
SI_EF_MAX_TILTED_SURFACE_LATITUDE
}{
Maximum Solar Irradiance for Equator Facing Latitude Tilt
}
\item{
SI_EF_MAX_TILTED_SURFACE
}{
Maximum Solar Irradiance for Equator Facing Tilted Surfaces (Set of
Surfaces)
}
\item{
SI_EF_MAX_TILTED_SURFACE_VERTICAL
}{
Maximum Solar Irradiance for Equator Facing Vertical Surface
}
\item{
MIDDAY_INSOL
}{
Midday Insolation Incident
}
\item{
MIDDAY_INSOL_MAX
}{
Midday Insolation Incident Maximum
}
\item{
MIDDAY_INSOL_MIN
}{
Midday Insolation Incident Minimum
}
\item{
MIDDAY_INSOL_SD
}{
Midday Insolation Incident Standard Deviation
}
\item{
INSOL_CONSEC_01_MIN
}{
Minimum Insolation Over A Consecutive 1 - day Period
}
\item{
INSOL_CONSEC_14_MIN
}{
Minimum Insolation Over A Consecutive 14 - day Period
}
\item{
INSOL_CONSEC_21_MIN
}{
Minimum Insolation Over A Consecutive 21 - day Period
}
\item{
INSOL_CONSEC_03_MIN
}{
Minimum Insolation Over A Consecutive 3 - day Period
}
\item{
INSOL_CONSEC_07_MIN
}{
Minimum Insolation Over A Consecutive 7 - day Period
}
\item{
INSOL_CONSEC_MONTH_MIN
}{
Minimum Insolation Over A Consecutive Month Period
}
\item{
SI_EF_MIN_TILTED_SURFACE_OPTIMAL
}{
Minimum Solar Irradiance Optimal
}
\item{
SI_EF_MIN_TILTED_SURFACE_OPTIMAL_ANG
}{
Minimum Solar Irradiance Optimal Angle
}
\item{
SI_EF_MIN_TILTED_SURFACE_OPTIMAL_ANG_ORT
}{
Minimum Solar Irradiance Tilted Surface Orientation
}
\item{
SI_EF_MIN_TILTED_SURFACE_HORIZONTAL
}{
Minimum Solar Irradiance for Equator Facing Horizontal Surface
}
\item{
SI_EF_MIN_TILTED_SURFACE_LAT_MINUS15
}{
Minimum Solar Irradiance for Equator Facing Latitude Minus 15 Tilt
}
\item{
SI_EF_MIN_TILTED_SURFACE_LAT_PLUS15
}{
Minimum Solar Irradiance for Equator Facing Latitude Plus 15 Tilt
}
\item{
SI_EF_MIN_TILTED_SURFACE_LATITUDE
}{
Minimum Solar Irradiance for Equator Facing Latitude Tilt
}
\item{
SI_EF_MIN_TILTED_SURFACE
}{
Minimum Solar Irradiance for Equator Facing Tilted Surfaces (Set of
Surfaces)
}
\item{
SI_EF_MIN_TILTED_SURFACE_VERTICAL
}{
Minimum Solar Irradiance for Equator Facing Vertical Surface
}
\item{
V10M
}{
Northward Wind at 10 Meters
}
\item{
V2M
}{
Northward Wind at 2 Meters
}
\item{
V50M
}{
Northward Wind at 50 Meters
}
\item{
PBLTOP
}{
Planetary Boundary Layer Top Pressure
}
\item{
PW
}{
Precipitable Water
}
\item{
PRECTOTCORR
}{
Precipitation Corrected
}
\item{
PRECTOTCORR_SUM
}{
Precipitation Corrected Sum
}
\item{
PRECSNOLAND
}{
Precipitation Land
}
\item{
GWETPROF
}{
Profile Soil Moisture
}
\item{
RH2M
}{
Relative Humidity at 2 Meters
}
\item{
GWETROOT
}{
Root Zone Soil Wetness
}
\item{
SLP
}{
Sea Level Pressure
}
\item{
SNODP
}{
Snow Depth
}
\item{
PRECSNO
}{
Snow Precipitation
}
\item{
PRECSNOLAND_SUM
}{
Snow Precipitation Land Sum
}
\item{
SG_SAA
}{
Solar Azimuth Angle
}
\item{
SG_NOON
}{
Solar Geometry Based Noon
}
\item{
SG_DAY_HOURS
}{
Solar Geometry Day Hours
}
\item{
SOLAR_DEFICITS_BLW_CONSEC_01
}{
Solar Irradiance Deficit Over A Consecutive 1 - day Period
}
\item{
SOLAR_DEFICITS_BLW_CONSEC_14
}{
Solar Irradiance Deficit Over A Consecutive 14 - day Period
}
\item{
SOLAR_DEFICITS_BLW_CONSEC_21
}{
Solar Irradiance Deficit Over A Consecutive 21 - day Period
}
\item{
SOLAR_DEFICITS_BLW_CONSEC_03
}{
Solar Irradiance Deficit Over A Consecutive 3 - day Period
}
\item{
SOLAR_DEFICITS_BLW_CONSEC_07
}{
Solar Irradiance Deficit Over A Consecutive 7 - day Period
}
\item{
SOLAR_DEFICITS_BLW_CONSEC_MONTH
}{
Solar Irradiance Deficit Over A Consecutive Month Period
}
\item{
SI_EF_MIN_TILTED_SURFACE_TRACKER
}{
Solar Irradiance Irradiance Tracking the Sun
}
\item{
SI_EF_TILTED_SURFACE_OPTIMAL
}{
Solar Irradiance Optimal
}
\item{
SI_EF_TILTED_SURFACE_OPTIMAL_ANG
}{
Solar Irradiance Optimal Angle
}
\item{
SI_EF_TILTED_SURFACE_OPTIMAL_ANG_ORT
}{
Solar Irradiance Tilted Surface Orientation
}
\item{
SI_EF_TILTED_SURFACE_TRACKER
}{
Solar Irradiance Tracking the Sun
}
\item{
SI_EF_TILTED_SURFACE_HORIZONTAL
}{
Solar Irradiance for Equator Facing Horizontal Surface
}
\item{
SI_EF_TILTED_SURFACE_LAT_MINUS15
}{
Solar Irradiance for Equator Facing Latitude Minus 15 Tilt
}
\item{
SI_EF_TILTED_SURFACE_LAT_PLUS15
}{
Solar Irradiance for Equator Facing Latitude Plus 15 Tilt
}
\item{
SI_EF_TILTED_SURFACE_LATITUDE
}{
Solar Irradiance for Equator Facing Latitude Tilt
}
\item{
SI_EF_TILTED_SURFACE
}{
Solar Irradiance for Equator Facing Tilted Surfaces (Set of Surfaces)
}
\item{
SI_EF_TILTED_SURFACE_VERTICAL
}{
Solar Irradiance for Equator Facing Vertical Surface
}
\item{
SG_SZA
}{
Solar Zenith Angle
}
\item{
SZA
}{
Solar Zenith Angle
}
\item{
QV10M
}{
Specific Humidity at 10 Meters
}
\item{
QV2M
}{
Specific Humidity at 2 Meters
}
\item{
RHOA
}{
Surface Air Density
}
\item{
SRF_ALB_ADJ
}{
Surface Albedo Adjusted
}
\item{
PS
}{
Surface Pressure
}
\item{
Z0M
}{
Surface Roughness
}
\item{
GWETTOP
}{
Surface Soil Wetness
}
\item{
INSOL_CONSEC_01
}{
Surplus Insolation Over A Consecutive 1 - day Period
}
\item{
INSOL_CONSEC_14
}{
Surplus Insolation Over A Consecutive 14 - day Period
}
\item{
INSOL_CONSEC_21
}{
Surplus Insolation Over A Consecutive 21 - day Period
}
\item{
INSOL_CONSEC_03
}{
Surplus Insolation Over A Consecutive 3 - day Period
}
\item{
INSOL_CONSEC_07
}{
Surplus Insolation Over A Consecutive 7 - day Period
}
\item{
INSOL_CONSEC_MONTH
}{
Surplus Insolation Over A Consecutive Month Period
}
\item{
T10M
}{
Temperature at 10 Meters
}
\item{
T10M_MAX
}{
Temperature at 10 Meters Maximum
}
\item{
T10M_MAX_AVG
}{
Temperature at 10 Meters Maximum Average
}
\item{
T10M_MIN
}{
Temperature at 10 Meters Minimum
}
\item{
T10M_MIN_AVG
}{
Temperature at 10 Meters Minimum Average
}
\item{
T10M_RANGE
}{
Temperature at 10 Meters Range
}
\item{
T10M_RANGE_AVG
}{
Temperature at 10 Meters Range Average
}
\item{
T2M
}{
Temperature at 2 Meters
}
\item{
T2M_MAX
}{
Temperature at 2 Meters Maximum
}
\item{
T2M_MAX_AVG
}{
Temperature at 2 Meters Maximum Average
}
\item{
T2M_MIN
}{
Temperature at 2 Meters Minimum
}
\item{
T2M_MIN_AVG
}{
Temperature at 2 Meters Minimum Average
}
\item{
T2M_RANGE
}{
Temperature at 2 Meters Range
}
\item{
T2M_RANGE_AVG
}{
Temperature at 2 Meters Range Average
}
\item{
TOA_SW_DNI
}{
Top - Of - Atmosphere Shortwave Direct Normal Radiation
}
\item{
TOA_SW_DNI_MAX
}{
Top - Of - Atmosphere Shortwave Direct Normal Radiation Maximum
}
\item{
TOA_SW_DNI_MIN
}{
Top - Of - Atmosphere Shortwave Direct Normal Radiation Minimum
}
\item{
TOA_SW_DNI_SD
}{
Top - Of - Atmosphere Shortwave Direct Normal Radiation Standard Deviation
}
\item{
TOA_SW_DWN
}{
Top - Of - Atmosphere Shortwave Downward Irradiance
}
\item{
TOA_SW_DWN_MAX
}{
Top - Of - Atmosphere Shortwave Downward Irradiance Maximum
}
\item{
TOA_SW_DWN_MIN
}{
Top - Of - Atmosphere Shortwave Downward Irradiance Minimum
}
\item{
TOA_SW_DWN_SD
}{
Top - Of - Atmosphere Shortwave Downward Irradiance Standard Deviation
}
\item{
TO3
}{
Total Column Ozone
}
\item{
TQV
}{
Total Column Precipitable Water
}
\item{
TROPPB
}{
Tropopause Pressure based on blended estimate.
}
\item{
TROPQ
}{
Tropopause Specific Humidity using blended TROPP estimate.
}
\item{
TROPT
}{
Tropopause Temperature using blended TROPP estimate.
}
\item{
T2MWET
}{
Wet Bulb Temperature at 2 Meters
}
\item{
WD10M
}{
Wind Direction at 10 Meters
}
\item{
WD2M
}{
Wind Direction at 2 Meters
}
\item{
WD50M
}{
Wind Direction at 50 Meters
}
\item{
WS10M
}{
Wind Speed at 10 Meters
}
\item{
WS10M_MAX
}{
Wind Speed at 10 Meters Maximum
}
\item{
WS10M_MAX_AVG
}{
Wind Speed at 10 Meters Maximum Average
}
\item{
WS10M_MIN
}{
Wind Speed at 10 Meters Minimum
}
\item{
WS10M_MIN_AVG
}{
Wind Speed at 10 Meters Minimum Average
}
\item{
WS10M_RANGE
}{
Wind Speed at 10 Meters Range
}
\item{
WS10M_RANGE_AVG
}{
Wind Speed at 10 Meters Range Average
}
\item{
WS2M
}{
Wind Speed at 2 Meters
}
\item{
WS2M_MAX
}{
Wind Speed at 2 Meters Maximum
}
\item{
WS2M_MAX_AVG
}{
Wind Speed at 2 Meters Maximum Average
}
\item{
WS2M_MIN
}{
Wind Speed at 2 Meters Minimum
}
\item{
WS2M_MIN_AVG
}{
Wind Speed at 2 Meters Minimum Average
}
\item{
WS2M_RANGE
}{
Wind Speed at 2 Meters Range
}
\item{
WS2M_RANGE_AVG
}{
Wind Speed at 2 Meters Range Average
}
\item{
WS50M
}{
Wind Speed at 50 Meters
}
\item{
WS50M_MAX
}{
Wind Speed at 50 Meters Maximum
}
\item{
WS50M_MAX_AVG
}{
Wind Speed at 50 Meters Maximum Average
}
\item{
WS50M_MIN
}{
Wind Speed at 50 Meters Minimum
}
\item{
WS50M_MIN_AVG
}{
Wind Speed at 50 Meters Minimum Average
}
\item{
WS50M_RANGE
}{
Wind Speed at 50 Meters Range
}
\item{
WS50M_RANGE_AVG
}{
Wind Speed at 50 Meters Range Average
}
\item{
ZENITH_LUMINANCE
}{
Zenith luminance
}
\item{
DISPH
}{
Zero Plane Displacement Height
}
}
}

\references{
\url{https://power.larc.nasa.gov/docs/}
}
\author{
Sparks, A. H. \email{adamhsparks@gmail.com}
}
\keyword{datasets}
