% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_groupings.R
\name{query_groupings}
\alias{query_groupings}
\title{Query the POWER API for Detailed Information on Available Parameter Groupings}
\usage{
query_groupings(global = FALSE)
}
\arguments{
\item{global}{Boolean; should the query return global parameter groupings and
attribute information? Defaults to \code{FALSE}, returning details for point data.}
}
\value{
A list object of information on parameter groupings in the POWER API.
}
\description{
Queries the POWER API returning detailed information on available parameter
groupings. Results are grouped by community followed by temporal API, or if
\code{global = TRUE}, grouped by climatology then by available parameter types.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Fetch groupings for parameters
query_groupings()

# Fetch groupings for global parameters
query_groupings(global = TRUE)
\dontshow{\}) # examplesIf}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
