% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parameters.R
\name{query_par}
\alias{query_par}
\title{Query the POWER API for detailed information on a given POWER parameter}
\usage{
query_par(par)
}
\arguments{
\item{par}{A character vector of a single solar, meteorological or
climatology parameter to query.  If unsure check \code{\link[=query_parameters]{query_parameters()}} for a
full list of parameters available for each temporal \acronym{API} and
community.}
}
\value{
A \link{list} defining the given \code{par} and providing valid temporal
\acronym{API} endpoints and communities for the given \code{par}.
}
\description{
Queries the \acronym{POWER} \acronym{API} returning detailed information on
a single user-provided parameter, \code{par}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
query_par("T2M")
\dontshow{\}) # examplesIf}
}
\seealso{
query_parameters
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
