% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parameters.R
\name{query_parameters}
\alias{query_parameters}
\title{Query the POWER API for Detailed Information on Available Parameters}
\usage{
query_parameters(
  community = NULL,
  pars = NULL,
  temporal_api = NULL,
  metadata = FALSE
)
}
\arguments{
\item{community}{An optional character vector providing community name:
"ag", "sb", or "re".}

\item{pars}{An optional character string of a single solar, meteorological,
or climatology parameter to query. If not provided, all parameters are
returned.}

\item{temporal_api}{An optional character vector indicating the temporal
API endpoint for data being queried. Supported values are "hourly",
"daily", "monthly", or "climatology".}

\item{metadata}{Boolean; retrieve extra parameter metadata? This is only
applicable if you supply both \code{community} and \code{temporal_api}; otherwise
it will be ignored. Defaults to \code{FALSE}.}
}
\value{
A list object of information for the requested parameter(s),
community(ies), and temporal API(s).
}
\description{
Queries the POWER API returning detailed information on available parameters.
For a list of all available parameters, see \code{parameters}.
}
\section{Argument details for \code{temporal_api}}{

There are four valid values:
\describe{
\item{hourly}{The hourly average of parameters by hour, day, month, and
year.}
\item{daily}{The daily average of parameters by day, month, and year.}
\item{monthly}{The monthly average of parameters by month and year.}
\item{climatology}{Parameters as 22-year climatologies (solar) and
30-year climatologies (meteorology); includes period climatology and
monthly average, maximum, and/or minimum values.}
}
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Fetch complete attribute information for "T2M"
query_parameters(pars = "T2M")

# Fetch temporal and community-specific attribute information
# for "T2M" in the "ag" community for the "hourly" temporal API
query_parameters(
  pars = "T2M",
  community = "ag",
  temporal_api = "hourly"
)

# Fetch all parameters in the "ag" community for "hourly" temporal API
query_parameters(
  community = "ag",
  temporal_api = "hourly"
)
\dontshow{\}) # examplesIf}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
