% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power.R
\name{.build_query}
\alias{.build_query}
\title{Construct a list of options to pass to the POWER API}
\usage{
.build_query(
  community,
  lonlat_identifier,
  pars,
  dates,
  site_elevation,
  wind_elevation,
  wind_surface,
  time_standard
)
}
\arguments{
\item{community}{A validated value for \code{community}.}

\item{lonlat_identifier}{A list of values, a result of \link{.check_lonlat}.}

\item{pars}{A validated value from \link{.check_pars}.}

\item{dates}{A list of values, a result of \link{.check_dates}..}

\item{site_elevation}{A validated value passed by \code{check_inputs}.}

\item{wind_elevation}{A validated value passed by \code{check_inputs}.}

\item{wind_surface}{A validated value passed by \code{check_inputs}.}
}
\value{
A \code{list} object of values to be passed to a \CRANpkg{crul} object to
query
the 'POWER' 'API'.
}
\description{
Construct a list of options to pass to the POWER API
}
\keyword{internal}
