% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.match_surface_alias}
\alias{.match_surface_alias}
\title{Match Wind Surface Aliases for Validity}
\usage{
.match_surface_alias(x)
}
\description{
Match Wind Surface Aliases for Validity
}
\keyword{internal}
