% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.send_mgmt_query}
\alias{.send_mgmt_query}
\title{Sends the query to the POWER management API}
\usage{
.send_mgmt_query(.url)
}
\arguments{
\item{.url}{A character string of the URL to be used for the \acronym{API}
query.}
}
\value{
A response from the POWER server containing either an error message
or the data requested.
}
\description{
There are no parameters that these API end points accept.
}
\keyword{internal}
