% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell.R
\name{get_cell}
\alias{get_cell}
\title{Download NASA-POWER Agroclimatology Variables for a Given Cell and Return a Tidy Data Frame}
\usage{
get_cell(lonlat = NULL, vars = c("T2M", "T2MN", "T2MX", "RH2M"),
  stdate = "1983-1-1", endate = Sys.Date())
}
\arguments{
\item{lonlat}{A length-2 numeric vector giving the decimal degree longitude
and latitude in that order for cell data to download}

\item{vars}{Weather variables to download, defaults to T2M, T2MN, T2MX and
RH2m. Valid variables are:
\itemize{
\item{\strong{toa_dwn} - Average top-of-atmosphere insolation (MJ/m^2/day)}
\item{\strong{swv_dwn} - Average insolation incident on a horizontal surface
(MJ/m^2/day)}
\item{\strong{lwv_dwn} - Average downward longwave radiative flux
(MJ/m^2/day)}
\item{\strong{T2M} - Average air temperature at 2m above the surface of the
Earth (degrees C)}
\item{\strong{T2MN} - Minimum air temperature at 2m above the surface of the
Earth (degrees C)}
\item{\strong{T2MX} - Maximum air temperature at 2m above the surface of the
Earth (degrees C)}
\item{\strong{RH2M} - Relative humidity at 2m above the surface of the Earth
(\%)}
\item{\strong{DFP2M} - Dew/Frost point temperature at 2m above the surface of
the Earth (degrees C)}
\item{\strong{RAIN} - Average precipitation (mm/day)}
\item{\strong{WS10M} - Wind speed at 10m above the surface of the Earth
(m/s)} }}

\item{stdate}{Starting date for download, defaults to 01/01/1983 (there is no
earlier data)}

\item{endate}{End date for download, defaults to current date}
}
\value{
A tidy \code{\link[base]{data.frame}} object of the requested variable(s)
for the requested longitude and latitude values.
}
\description{
Download NASA-POWER (Prediction of Worldwide Energy Resource)
agroclimatology data for a given 1 degree x 1 degree cell for a range of
years.

The agroclimatology data are satellite and modelled derived solar and
meteorological data. The POWER website describes the data as:
\itemize{
  \item{Near real-time 1째 x 1째 daily time series of solar radiation and
  meteorology}
  \item{Daily total solar radiation from July 1, 1983 through near real-time}
  \item{Daily averaged air temperature (average/minimum/maximum/dew point)
  from January 1, 1983 through near real-time}
  \item{Daily averaged precipitation from January 1997 through February 2013}
  \item{Global coverage on a 1째 latitude by 1째 longitude grid}
}
}
\examples{
\dontrun{
nasa <- get_cell(lonlat = c(-179.5, -89.5))
}

}
\references{
\url{https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf}
}
\author{
Adam H. Sparks, adamhsparks@gmail.com
}
