% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasapower-package.R
\docType{package}
\name{nasapower}
\alias{nasapower}
\alias{nasapower-package}
\title{Downloads NASA-POWER Agroclimatology Variables and Returns a Tidy Data Frame}
\description{
Downloads NASA-POWER Agroclimatology Variables and Returns a Tidy Data Frame
}
\seealso{
\link{get_cell}
\link{get_region}
}
\author{
Adam H Sparks
}
