% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasapower-package.R
\docType{package}
\name{nasapower}
\alias{nasapower}
\alias{nasapower-package}
\title{NASA-POWER Agroclimatology Data from R}
\description{
\code{nasapower} provides an R interface to download NASA-POWER
agroclimatology data and create a tidy data frame for individual cells or a
specified region.  POWER (Prediction Of Worldwide Energy Resource) data are
freely available for download through a web interface at a resolution of 1
degree longitude by 1 degree latitude funded through the NASA Earth Science
Directorate Applied Science Program.
}
\note{
While \code{nasapower} does not redistribute the data in anyway, we
encourage users to follow the requests of the POWER Project Team.

\preformatted{
When POWER data products are used in a publication, we request the
following acknowledgment be included:

"These data were obtained from the NASA Langley Research Center
POWER Project funded through the NASA Earth Science Directorate Applied
Science Program."
}
}
\references{
\url{https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf}
}
\seealso{
\itemize{
\item \link{get_cell} Download NASA-POWER Agroclimatology Variables for a
Given Cell and Return a Tidy Data Frame
\item \link{get_region} Download NASA-POWER Agroclimatology Variables for a
Given Region and Return a Tidy Data Frame
}
}
\author{
Adam H Sparks

Adam H. Sparks, \email{adamhsparks@gmail.com}
}
