% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{NASA - POWER Parameters Available for Download}
\format{A list with 141 objects}
\source{
\url{https://power.larc.nasa.gov/RADAPP/GEODATA/powerWeb/POWER_Parameters_v108.json}
}
\usage{
parameters
}
\description{
An \pkg{R} list object of metadata for NASA-POWER parameters available for
querying from the NASA-POWER database.
}
\references{
\url{https://power.larc.nasa.gov/docs/v1/}
}
\author{
Adam H Sparks \email{adamhsparks@gmail.com}
}
\keyword{datasets}
