% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{NASA POWER parameters available for download}
\format{
A list with 389 weather and climate parameters contained within the
database.
}
\source{
\url{https://power.larc.nasa.gov/beta/api/system/manager/system/parameters}
}
\usage{
parameters
}
\description{
An \R \code{list} object of \acronym{POWER} parameters and metadata available for
querying from the \acronym{POWER} database.
}
\references{
\url{https://power.larc.nasa.gov/docs/}
}
\author{
Sparks, A. H. \email{adamhsparks@gmail.com}
}
\keyword{datasets}
