% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{print.POWER.Info}
\alias{print.POWER.Info}
\title{Prints Power.Info object.}
\usage{
\method{print}{POWER.Info}(x, ...)
}
\arguments{
\item{x}{POWER.Info object}

\item{...}{ignored}
}
\description{
Prints Power.Info object.
}
