% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parameters.R
\name{query_parameters}
\alias{query_parameters}
\title{Query the POWER API for parameter descriptions}
\usage{
query_parameters(temporal_api, community)
}
\arguments{
\item{temporal_api}{Temporal average for data being queried, supported
values are \dQuote{hourly}, \dQuote{daily}, \dQuote{monthly} or
\dQuote{climatology}.  See argument details for more.}

\item{community}{A character vector providing community name: \dQuote{ag},
\dQuote{sb} or \dQuote{re}.  See argument details for more.}
}
\value{
A \link{list} of definitions providing valid parameters and their
definitions for each of the three \acronym{API} endpoints and three
communities
}
\description{
Queries the \acronym{POWER} \acronym{API} for detailed information for all
parameters provided by the user-provided community and temporal \acronym{API}
endpoints.
}
\section{Argument details for \dQuote{community}}{
 there are three valid
values, one must be supplied.  This  will affect the units of the parameter
and the temporal display of time series data.

\describe{
\item{ag}{Provides access to the Agroclimatology Archive, which
contains industry-friendly parameters formatted for input to crop models.}

\item{sb}{Provides access to the Sustainable Buildings Archive, which
contains industry-friendly parameters for the buildings community to include
parameters in multi-year monthly averages.}

\item{re}{Provides access to the Renewable Energy Archive, which contains
parameters specifically tailored to assist in the design of solar and wind
powered renewable energy systems.}
}
}

\section{Argument details for \code{temporal_average}}{
 There are four valid
values.
\describe{
\item{hourly}{The hourly average of \code{pars} by hour, day, month and
year.}
\item{daily}{The daily average of \code{pars} by day, month and year.}
\item{monthly}{The monthly average of \code{pars} by month and year.}
\item{climatology}{Provide parameters as 22-year climatologies (solar)
and 30-year climatologies (meteorology); the period climatology and
monthly average, maximum, and/or minimum values.}
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get a full list of all valid parameters for the 'DAILY' API for the 'AG'
# community
query_parameters(temporal_api = "DAILY", community = "AG")
\dontshow{\}) # examplesIf}
}
\seealso{
query_par
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
